Pass all intervals in scope so we can produce increments between them.
We don't want to use -I as it will pick up unwanted intermediate
snapshots - for example, the 'frequent' snapshots between two hours,
or the 'daily' snapshot between 00h00 and 01h00. Instead we look
through all the snapshots in order and skip the ones that we don't
want shipped to the remote system, then produce incrementals between
successive pairs.